// src/modules/article/types/article.types.ts
export type ARTICLE_EVENT =
  | "PUBLISH_ARTICLE"
  | "SUBMIT_ARTICLE"
  | "REJECT_ARTICLE"
  | "ARCHIVE_ARTICLE"
  | "RESUBMIT_ARTICLE"
  | "DRAFT_ARTICLE"
  | "REVIEW_ARTICLE";

export const ARTICLE_EVENT_LIST = [
  "PUBLISH_ARTICLE",
  "REJECT_ARTICLE",
  "ARCHIVE_ARTICLE",
  "RESUBMIT_ARTICLE",
  "DRAFT_ARTICLE",
  "REVIEW_ARTICLE",
] as ARTICLE_EVENT[];
export type ARTICLE_STATUS =
  | "DRAFTED"
  | "PUBLISHED"
  | "PENDING_REVIEW"
  | "REJECTED"
  | "ARCHIVED"
  | "RESUBMITTED";

export const ARTICLE_STATUS_LIST = [
  "DRAFTED",
  "PUBLISHED",
  "PENDING_REVIEW",
  "REJECTED",
  "ARCHIVED",
  "RESUBMITTED",
];